requirements.txt
asgiref==3.2.7
astroid==2.3.3
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.5
django-ckeditor==5.9.0
django-heroku==0.3.1
django-html-field==0.1.3
django-instagram==0.3.1
django-js-asset==1.2.2
django-otp==0.8.1
django-recaptcha==2.0.6
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.9.5
jsmin==2.2.2
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
Pillow==7.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.4
pylint==2.4.4
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.12.0
pytz==2019.3
qrcode==6.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
